Finished the review of the py-examples - Not tested
