+++ examples/example_GaussianFieldsCosMethod.py
