+++ examples/example_single_InterferogramMPI.py
@@ -1,26 +1,30 @@
-'''script to do synthetic interferogram with pvti files
+'''
+Script to do synthetic interferogram with pvti files
+
+Author: Louis Evans
+Reviewer: Stefano Merlini
 
 Run with the following job script:
+	#!/bin/sh
+	#PBS -l walltime=HH:MM:SS
+	#PBS -l select=1:ncpus=N:mpiprocs=N:mem=Mgb
+	#PBS -j oe
+	cd '/rds/general/user/le322/home/synthPy'
 
-#!/bin/sh
-#PBS -l walltime=HH:MM:SS
-#PBS -l select=1:ncpus=N:mpiprocs=N:mem=Mgb
-#PBS -j oe
-cd '/rds/general/user/le322/home/synthPy'
+	module load anaconda3/personal
 
-module load anaconda3/personal
+	source activate MAGPIE_venv #load venv
 
-source activate MAGPIE_venv #load venv
-
-mpiexec  -n <n_cpus> python run_scripts/interference_MPI.py <number of rays> <path/to/pvti> <output directory>
+	mpiexec  -n <n_cpus> python run_scripts/interference_MPI.py <number of rays> <path/to/pvti> <output directory>
 '''
 
 import sys
-sys.path.insert(1, '/rds/general/user/le322/home/synthPy')
+sys.path.append('../synthPy/')      # import path/to/synthpy
 import numpy as np
 from mpi4py import MPI
 import pickle
 from mpi4py.util import pkl5
+import utils.handle_filetypes as utilIO
 import solver.full_solver as s
 import solver.rtm_solver as rtm
 import matplotlib.pyplot as plt
@@ -34,38 +38,6 @@ rank = comm.Get_rank()
 Np_ray_split = int(5e5)
 num_processors = comm.Get_size()
 
-def pvti_readin(filename):
-	'''
-	Reads in data from pvti with filename, use this to read in electron number density data
-
-	'''
-
-	reader = vtk.vtkXMLPImageDataReader()
-	reader.SetFileName(filename)
-	reader.Update()
-
-	data = reader.GetOutput()
-	dim = data.GetDimensions()
-	spacing = np.array(data.GetSpacing())
-
-	v = vtk_np.vtk_to_numpy(data.GetCellData().GetArray(0))
-	n_comp = data.GetCellData().GetArray(0).GetNumberOfComponents()
-	
-	vec = [int(i-1) for i in dim]
-
-	if(n_comp > 1):
-		vec.append(n_comp)
-
-	if(n_comp > 2):
-		img = v.reshape(vec,order="F")[0:dim[0]-1,0:dim[1]-1,0:dim[2]-1,:]
-	else:
-		img = v.reshape(vec,order="F")[0:dim[0]-1,0:dim[1]-1,0:dim[2]-1]
-
-	dim = img.shape
-
-	return img,dim,spacing
-
-
 if __name__ == '__main__':
 	#retrieve input variables
 	Np = int(float(sys.argv[1]))
@@ -73,7 +45,7 @@ if __name__ == '__main__':
 	output_loc = str(sys.argv[3])
 
 	#load pvti 
-	ne, dim, spacing = pvti_readin(str(file_loc))
+	ne, dim, spacing = utilIO.pvti_readin(str(file_loc))
 	extent_x = ((dim[0]*spacing[0])/2)
 	extent_y = ((dim[1]*spacing[1])/2)
 	extent_z = ((dim[2]*spacing[2])/2)
@@ -101,16 +73,7 @@ if __name__ == '__main__':
 			file_loc is {file_loc}
 			wl is {wl}
 			''')
-		print('''    
-	.______          ___   ____    ____    .___________..______          ___       ______  __  .__   __.   _______ 
-	|   _  \        /   \  \   \  /   /    |           ||   _  \        /   \     /      ||  | |  \ |  |  /  _____|
-	|  |_)  |      /  ^  \  \   \/   /     `---|  |----`|  |_)  |      /  ^  \   |  ,----'|  | |   \|  | |  |  __  
-	|      /      /  /_\  \  \_    _/          |  |     |      /      /  /_\  \  |  |     |  | |  . `  | |  | |_ | 
-	|  |\  \----./  _____  \   |  |            |  |     |  |\  \----./  _____  \ |  `----.|  | |  |\   | |  |__| | 
-	| _| `._____/__/     \__\  |__|            |__|     | _| `._____/__/     \__\ \______||__| |__| \__|  \______| 
-																												
-	''')
-
+		print('''Ray-Tracing...''')
 
 	def system_solve(Np,beam_size,divergence,field, ne_extent, probing_direction, wl):
 		## Initialise laser beam
@@ -126,7 +89,6 @@ if __name__ == '__main__':
 		interferogram =rtm.InterferometerRays(rf, E = E)
 		interferogram.solve(wl = wl)
 		interferogram.interferogram(bin_scale = 1, clear_mem=True)
-
 		return interferogram
 
 	# split ray bundle
